feat: Add password visibility toggle to ResetPass component
This commit introduces a password visibility toggle feature to the ResetPass component. The user can now choose to show or hide the password while typing in both the \"New Password\" and \"Confirm New Password\" fields. This is achieved by adding a button with an eye icon next to each password field, which toggles the `type` attribute of the input field between \"password\" and \"text\".